requirements.txt
langchain-core
langchain-openai
langchain-community
python-dotenv
langchain-google-genai
faiss-gpu